feat(fixed periods): add getPeriodByDate util (works with month only)
