+++ src/period-calculation/month-numbers.ts
@@ -1,0 +1,16 @@
+const monthNumbers = {
+    JAN: { value: 1, name: 'January' },
+    FEB: { value: 2, name: 'February' },
+    MAR: { value: 3, name: 'March' },
+    APR: { value: 4, name: 'April' },
+    MAY: { value: 5, name: 'May' },
+    JUN: { value: 6, name: 'June' },
+    JUL: { value: 7, name: 'July' },
+    AUG: { value: 8, name: 'August' },
+    SEP: { value: 9, name: 'September' },
+    OCT: { value: 10, name: 'October' },
+    NOV: { value: 11, name: 'November' },
+    DEC: { value: 12, name: 'December' },
+}
+
+export default monthNumbers
