+++ src/period-calculation/monthly-periods/build-label.ts
@@ -1,0 +1,84 @@
+import { Temporal } from '@js-temporal/polyfill'
+import { SupportedCalendar } from '../../types'
+import { isCustomCalendar } from '../../utils/helpers'
+import localisationHelpers from '../../utils/localisationHelpers'
+import { MULTI_MONTH_PERIOD_TYPES } from '../period-types'
+import { PeriodIdentifier } from '../types'
+
+type BuildLabelFunc = (options: {
+    periodType: PeriodIdentifier
+    month: Temporal.PlainDate
+    nextMonth: Temporal.PlainDate
+    index: number
+    locale: string
+    calendar: SupportedCalendar
+}) => string
+
+const buildLabel: BuildLabelFunc = (options) => {
+    const { periodType, month, nextMonth, calendar, locale } = options
+
+    if (isCustomCalendar(calendar)) {
+        return buildLabelForCustomCalendar(options)
+    }
+
+    const withYearFormat = {
+        month: 'long' as const,
+        year: 'numeric' as const,
+        calendar,
+    }
+    const monthOnlyFormat = {
+        month: 'long' as const,
+        calendar,
+    }
+
+    let result = ''
+
+    if (MULTI_MONTH_PERIOD_TYPES.includes(periodType)) {
+        const format =
+            month.year === nextMonth.year ? monthOnlyFormat : withYearFormat
+        result = `${month.toLocaleString(
+            locale,
+            format
+        )} - ${nextMonth.toLocaleString(locale, withYearFormat)}`
+    } else {
+        result = `${month.toLocaleString(locale, withYearFormat)}`
+    }
+
+    // needed for ethiopic calendar - the default formatter adds the era, which is not what we want in DHIS2
+    result = result.replace(/ERA\d+\s*/g, '').trim()
+    return result
+}
+
+const buildLabelForCustomCalendar: BuildLabelFunc = ({
+    periodType,
+    month,
+    nextMonth,
+    calendar,
+    locale,
+}) => {
+    let result = ''
+
+    if (MULTI_MONTH_PERIOD_TYPES.includes(periodType)) {
+        const showYear = month.year !== nextMonth.year
+        result = `${localisationHelpers.localiseMonth(
+            month,
+            { locale, calendar },
+            {}
+        )}${
+            showYear ? ` ${month.year}` : ''
+        } - ${localisationHelpers.localiseMonth(
+            nextMonth,
+            { locale, calendar },
+            {}
+        )} ${nextMonth.year}`
+    } else {
+        result = `${localisationHelpers.localiseMonth(
+            month,
+            { locale, calendar },
+            {}
+        )} ${nextMonth.year}`
+    }
+    return result
+}
+
+export default buildLabel
